[Refactor] PR대비 코드 정리
