--- keewe-api/src/main/java/ccc/keeweapi/api/TestController.java
@@ -1,15 +1,0 @@
-package ccc.keeweapi.api;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/test")
-public class TestController {
-
-    @GetMapping
-    public String test() {
-        return "test";
-    }
-}
