--- keewe-api/src/main/java/ccc/keeweapi/api/user/UserController.java
@@ -1,19 +1,0 @@
-package ccc.keeweapi.api.user;
-
-import ccc.keeweapi.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/user")
-@RequiredArgsConstructor
-public class UserController {
-    private final UserService userService;
-
-    @GetMapping
-    String getToken() {
-        return userService.getToken();
-    }
-}
