+++ keewe-api/src/main/java/ccc/keeweapi/config/security/SecurityConfig.java
@@ -1,5 +1,8 @@
 package ccc.keeweapi.config.security;
 
+import ccc.keeweapi.config.security.jwt.JwtAuthenticationEntryPoint;
+import ccc.keeweapi.config.security.jwt.JwtAuthenticationFilter;
+import ccc.keeweapi.config.security.jwt.JwtUtils;
 import ccc.keeweapi.service.UserPrincipalDetailsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
